fix: contract calls with ledger
